Removed last of old website requirements
"Hatsune Miku and VOCALOID: The History Behind the DIVA" was initially made as a school project for both my English class (as a podcast) and for my Web Design class in the form of this website. Over the last couple of months, I have decided that some of these old requirements, such as a full-on description of myself on the "About Me" page, the Table of Pages, and the Carousel on the index page, were unnecessary, so I have removed them. I appreciate your understanding.

I have also updated the README.md!